int max -> double max

+++ lmp9/trunk/src/gauss.c
@@ -8,7 +8,7 @@
  */
 int diag_elem_choice(Matrix *mat, Matrix *b, int j) {
     // int j is the current row
-    int max = mat->data[j][j];
+    double max = mat->data[j][j];
     int row = j;
     for (int i = j+1; i < mat->r; i++) {
         if (fabs(max) < fabs(mat->data[i][j])) {
